move order request models to common/models
