--- alpaca/trading/models/__init__.py
@@ -1,1 +1,0 @@
-from .requests import *
